clean up tests, and use a specific port to avoid collision with CI server.
