+++ Sources/DotNet/WoopsaTest/UnitTestWoopsaObjectAdapter.cs
@@ -6,126 +6,12 @@ using System.Collections.Generic;
 
 namespace WoopsaTest
 {
-    public enum Day { Monday, Thursday, Wednesday }
-
-    public class ClassAInner1Inner
-    {
-        public string APropertyString { get; set; }
-
-        [WoopsaVisible(true)]
-        public string APropertyVisible { get; set; }
-
-        [WoopsaVisible(false)]
-        public string APropertyHidden { get; set; }
-
-    }
-
-    public class ClassAInner1
-    {
-        public ClassAInner1()
-        {
-            Inner = new ClassAInner1Inner();
-        }
-
-        public int APropertyInt { get; set; }
-
-        [WoopsaVisible(false)]
-        public int APropertyIntHidden { get; set; }
-
-        [WoopsaVisible(true)]
-        public int APropertyIntVisible { get; set; }
-
-        [WoopsaVisible(true)]
-        public ClassAInner1Inner Inner { get; set; }
-
-    }
-
-
-    public class SubClassAInner1 : ClassAInner1
-    {
-        [WoopsaVisible(true)]
-        public int ExtraProperty { get; set; }
-
-    }
-
-    [WoopsaVisibility(WoopsaVisibility.DefaultIsVisible | WoopsaVisibility.Inherited | WoopsaVisibility.ObjectClassMembers)]
-    public class SubClassAInner2 : SubClassAInner1
-    {
-
-    }
-
-
-    public class ClassA
-    {
-        public ClassA()
-        {
-            Inner1 = new ClassAInner1();
-        }
-        public bool APropertyBool { get; set; }
-
-        [WoopsaVisible(false)]
-        public DateTime APropertyDateTime { get; set; }
-        public DateTime APropertyDateTime2 { get; set; }
-
-        [WoopsaVisible(true)]
-        public ClassAInner1 Inner1 { get; set; }
-
-        public Day Day { get; set; }
-
-    }
-
-    [WoopsaVisibility(WoopsaVisibility.DefaultIsVisible | WoopsaVisibility.Inherited | WoopsaVisibility.MethodSpecialName)]
-    public class ClassB : ClassA
-    {
-        public int APropertyInt { get; set; }
-
-        double APropertyDouble { get; set; }
-    }
-
-    [WoopsaVisibility(WoopsaVisibility.None)]
-    public class ClassC : ClassB
-    {
-        [WoopsaVisible]
-        [WoopsaValueType(WoopsaValueType.Integer)]
-        public string APropertyText { get; set; }
-
-        [WoopsaVisible]
-        public TimeSpan APropertyTimeSpan { get; set; }
-
-        // This one must not be published (private)
-        [WoopsaVisible]
-        private double APropertyDouble2 { get; set; }
-
-        [WoopsaVisible]
-        [WoopsaValueType(WoopsaValueType.JsonData)]
-        public string APropertyJson { get; set; }
-
-
-    }
-
-    public class ClassD
-    {
-        public ClassD(int n) { APropertyInt = n; }
-        public int APropertyInt { get; set; }
-    }
-
-    public class ClassE
-    {
-        public ClassE()
-        {
-        }
-
-        public Day Day { get; set; }
-
-    }
-
     [TestClass]
     public class UnitTestWoopsaObjectAdapter
     {
         [TestMethod]
         public void TestWoopsaObjectAdapter()
         {
-
             ClassA a = new ClassA();
             WoopsaObjectAdapter adapterA1 = new WoopsaObjectAdapter(null, "a", a);
             Assert.IsNotNull(adapterA1.Properties.ByNameOrNull(nameof(a.APropertyBool)));
@@ -463,37 +349,7 @@ namespace WoopsaTest
                 WoopsaVisibility.ListClassMembers);
             Assert.AreEqual(adapterList2.Items.Count, 1); // List "Item" indexer
         }
-
-        public interface InterfaceA
-        {
-            string A { get; set; }
-
-            void MethodA();
-        }
-
-        public interface InterfaceB : InterfaceA
-        {
-            int B { get; set; }
-
-            void MethodB();
-        }
-
-        public class ClassInterfaceA : InterfaceA
-        {
-            public string A { get; set; }
-
-            public void MethodA() { }
-        }
-
-        public class ClassInterfaceB : InterfaceB
-        {
-            public string A { get; set; }
-            public int B { get; set; }
-
-            public void MethodA() { }
-            public void MethodB() { }
-        }
-
+       
         [TestMethod]
         public void TestWoopsaObjectAdapterInterfaces()
         {
@@ -518,7 +374,5 @@ namespace WoopsaTest
             Assert.IsNotNull(adapterB2.Properties.ByNameOrNull(nameof(InterfaceA.A)));
             Assert.IsNotNull(adapterB2.Methods.ByNameOrNull(nameof(InterfaceA.MethodA)));
         }
-
-
     }
 }
